New State Machines
I made these state machines, however they depend on setting a variable to a function so I am not sure if they will work. We need current measurements from the intakes in order for any of this to work. I may have accidentally broken some other stuff, we will have to fix